refactor: home page
