+++ frontend/src/app/components/README.md
@@ -1,0 +1,5 @@
+# Avis subjectif de Nans
+
+Il vaut mieux mettre tous les composants partagés dans `src/components/`.
+
+Puis pour chaque `app/../page.tsx` créer un dossier `components/` pour les composants spécifiques à cette page.
